refactor: ♻️ rename "anagram" module to "anagrams"
